Move overlap visualization code into separate script
